pyproject.toml
[project]
name = "paid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Claudia Vanea", email = "claudia.vanea@hotmail.co.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "deepgram-sdk>=3.11.0",
    "elevenlabs>=1.56.1",
    "peewee>=3.17.9",
    "pyaudio>=0.2.14",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "streamlit>=1.44.1",
]

[project.scripts]
paid = "paid:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
